+++ biorxiv/journal_tracker/journal_recommendation_experiment.py
@@ -17,55 +17,37 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 from sklearn.dummy import DummyClassifier
-from sklearn.model_selection import GridSearchCV, KFold
 from sklearn.neighbors import KNeighborsClassifier
-
 from tqdm import tqdm_notebook
 
-
-# In[2]:
-
-
-cluster = not True
+from annorxiver_modules.journal_rec_helper import (
+    cross_validation, dummy_evaluate,
+    knn_evaluate, knn_centroid_evaluate
+)
 
 
 # # Load bioRxiv Document Vectors
 
-# In[3]:
+# In[2]:
 
 
-if cluster:
-    biorxiv_journal_df = (
-        pd.read_csv("mapped_published_doi.tsv", sep="\t")
-        .groupby("doi")
-        .agg({
-            "document":"last",
-            "category":"first",
-            "journal":"first",
-            "doi":"last",
-            "published_doi":"first",  
-            "pmcid":"first", 
-        })
-        .reset_index(drop=True)
-    )
-else:
-    biorxiv_journal_df = (
-        pd.read_csv("output/mapped_published_doi.tsv", sep="\t")
-        .groupby("doi")
-        .agg({
-            "document":"last",
-            "category":"first",
-            "journal":"first",
-            "doi":"last",
-            "published_doi":"first",  
-            "pmcid":"first", 
-        })
-        .reset_index(drop=True)
-    )
+biorxiv_journal_df = (
+    pd.read_csv("output/mapped_published_doi.tsv", sep="\t")
+    .groupby("doi")
+    .agg({
+        "document":"last",
+        "category":"first",
+        "journal":"first",
+        "doi":"last",
+        "published_doi":"first",  
+        "pmcid":"first", 
+    })
+    .reset_index(drop=True)
+)
 biorxiv_journal_df.head()
 
 
-# In[4]:
+# In[3]:
 
 
 # Count number of Non-NaN elements
@@ -74,7 +56,7 @@ print(f"Total number of entries: {biorxiv_journal_df.shape[0]}")
 print(f"Percent Covered: {biorxiv_journal_df.pmcid.count()/biorxiv_journal_df.shape[0]:.2f}")
 
 
-# In[5]:
+# In[4]:
 
 
 golden_set_df = biorxiv_journal_df.query("pmcid.notnull()")
@@ -83,37 +65,29 @@ golden_set_df.head()
 
 # # Load Pubmed Central Document Vectors
 
+# In[5]:
+
+
+pmc_articles_df = (
+    pd.read_csv(
+        "../../pmc/exploratory_data_analysis/output/pubmed_central_journal_paper_map.tsv.xz", 
+        sep="\t"
+    )
+    .query("article_type=='research-article'")
+)
+print(pmc_articles_df.shape)
+pmc_articles_df.head()
+
+
 # In[6]:
 
 
-if cluster:
-    pmc_articles_df = (
-        pd.read_csv(
-            "pubmed_central_journal_paper_map.tsv.xz", 
-            sep="\t"
-        )
-        .query("article_type=='research-article'")
-    )
-else:
-    pmc_articles_df = (
-        pd.read_csv(
-            "../../pmc/exploratory_data_analysis/output/pubmed_central_journal_paper_map.tsv.xz", 
-            sep="\t"
-        )
-        .query("article_type=='research-article'")
-    )
-print(pmc_articles_df.head())
-
-
-# In[7]:
-
-
-print(pmc_articles_df.journal.value_counts().shape)
 journals = pmc_articles_df.journal.value_counts()
+print(journals.shape)
 journals
 
 
-# In[8]:
+# In[7]:
 
 
 # Filter out low count journals
@@ -122,29 +96,20 @@ print(pmc_articles_df.shape)
 pmc_articles_df.head()
 
 
-# In[9]:
+# In[8]:
 
 
-if cluster:
-    pmc_embedding_dict = {
-        int(path.stem[-7:-4]):pd.read_csv(
-            str(path), 
-            sep="\t"
-        )
-        for path in Path("pmc_vectors/").rglob("*tsv.xz")
-    }
-else:
-    pmc_embedding_dict = {
-        int(path.stem[-7:-4]):pd.read_csv(
-            str(path), 
-            sep="\t"
-        )
-        for path in Path("../../pmc/word_vector_experiment/output/").rglob("*tsv.xz")
-    }
+pmc_embedding_dict = {
+    int(path.stem[-7:-4]):pd.read_csv(
+        str(path), 
+        sep="\t"
+    )
+    for path in Path("../../pmc/word_vector_experiment/output/").rglob("*tsv.xz")
+}
 pmc_embedding_dict[300].head()
 
 
-# In[10]:
+# In[9]:
 
 
 full_training_dataset = {
@@ -160,7 +125,7 @@ full_training_dataset = {
 }
 
 
-# In[11]:
+# In[10]:
 
 
 subsampled_training_dataset = {
@@ -180,91 +145,23 @@ subsampled_training_dataset = {
 
 # # Train Recommendation System
 
-# In[12]:
+# In[11]:
 
 
-def cross_validation(model, dataset, evaluate, cv=10, random_state=100, **kwargs):
-    
-    folds = KFold(n_splits=cv, random_state = random_state, shuffle=True)
-    cv_fold_accs = []
-    
-    fold_predictions = []
-    for train, val in folds.split(dataset):
-        
-        prediction, true_labels = evaluate(
-            model, dataset.iloc[train], 
-            dataset.iloc[val], **kwargs
-        )
-
-        accs = [
-                 (
-                     1 if true_labels[data_idx] in prediction_row 
-                     else 0 
-                 )
-                 for data_idx, prediction_row in enumerate(prediction)
-        ]
-        
-        cv_fold_accs.append(np.sum(accs)/len(accs))
-        print(f"{np.sum(accs)} out of {len(accs)}")
-        
-        fold_predictions.append(prediction)
-        
-    print(f"Total Accuracy: {np.mean(cv_fold_accs)*100:.3f}%")
-    return fold_predictions
+knn_model = KNeighborsClassifier(n_neighbors=10)
 
 
 # ## Random Journal Prediction
 
 # The central idea here is to answer the question what is the accuracy when journals are recommended at random?
 
-# In[13]:
-
-
-def dummy_evaluate(model, training_data, validation_data, **kwargs):
-    top_X = kwargs.get("top_predictions", 10)
-    random_states = kwargs.get("dummy_seed", [100,200,300,400,500,600,700,800,900,1000])
-    
-    X_train = (
-        training_data
-        .drop("journal", axis=1)
-        .values
-        .astype('float32')
-    )
-    
-    Y_train = (
-        training_data
-        .journal
-        .values
-    )
-    
-    X_val = (
-        validation_data
-        .drop("journal", axis=1)
-        .values
-        .astype('float32')
-    )
-    Y_val = (
-        validation_data
-        .journal
-        .values
-    )
-    
-    predictions = []
-    for i, seed in zip(range(top_X), random_states):
-        model.random_state = seed
-        model.fit(X_train, Y_train)
-        predictions.append(model.predict(X_val))
-
-    return np.stack(predictions).transpose(), Y_val
-
-
-# In[16]:
+# In[12]:
 
 
 model = DummyClassifier(strategy='uniform')
 
 
-# In[17]:
+# In[14]:
 
 
 _ = cross_validation(
@@ -277,54 +174,6 @@ _ = cross_validation(
 # ## KNearestNeighbors Paper by Paper Comparison - Full Dataset
 
 # Assuming I didn't take mega-journal influence into account, what would the initial recommendation accuracy be?
-
-# In[14]:
-
-
-def knn_evaluate(model, training_data, validation_data, **kwargs):
-    
-    X_train = (
-        training_data
-        .drop("journal", axis=1)
-        .values
-        .astype('float32')
-    )
-    
-    Y_train = (
-        training_data
-        .journal
-        .values
-    )
-    
-    X_val = (
-        validation_data
-        .drop("journal", axis=1)
-        .values
-        .astype('float32')
-    )
-    
-    Y_val = (
-        validation_data
-        .journal
-        .values
-    )
-    
-    model.fit(X_train, Y_train)
-    distance, neighbors = model.kneighbors(X_val)
-    
-    predictions = [
-        Y_train[neighbor_predict]
-        for neighbor_predict in neighbors 
-    ]
-
-    return np.stack(predictions), Y_val
-
-
-# In[18]:
-
-
-knn_model = KNeighborsClassifier(n_neighbors=10)
-
 
 # In[ ]:
 
@@ -339,12 +188,6 @@ _ = cross_validation(
 # ## KNearestNeighbors Paper by Paper Comparison Subsampled
 
 # The first idea for a classifier is to compare which papers are similar to other papers. Due to the overflow of PLOS One papers I sub-sampled each journal to have only 100 papers for representation. Then trained a KNearestNeighbors to determine how often does the correct journal appear in the top ten neighbors as well as top twenty neighbors.
-
-# In[16]:
-
-
-knn_model = KNeighborsClassifier(n_neighbors=10)
-
 
 # In[17]:
 
@@ -368,63 +211,7 @@ for dim in subsampled_training_dataset:
 
 # Following up on the original idea, I thought a helpful experiment would be to perform a centroid analysis (i.e. take the average of all papers within each journal). Similar to above I trained a KNearestNeighbors classifier to see if the correct journal will appear in the top 10 neighbors.
 
-# In[15]:
-
-
-def knn_centroid_evaluate(model, training_data, validation_data, **kwargs):
-    
-    train_centroid_df = (
-        training_data
-        .groupby("journal")
-        .agg("mean")
-        .reset_index()
-    )
-            
-    X_train_centroid = (
-        train_centroid_df
-        .drop("journal", axis=1)
-        .values
-        .astype('float32')
-    )
-
-    Y_train_centroid = (
-        train_centroid_df
-        .journal
-        .values
-    )
-    
-    
-    X_val = (
-        validation_data
-        .drop("journal", axis=1)
-        .values
-        .astype('float32')
-    )
-    
-    Y_val = (
-        validation_data
-        .journal
-        .values
-    )
-    
-    knn_model.fit(X_train_centroid, Y_train_centroid)
-    distance, neighbors = knn_model.kneighbors(X_val)
-    
-    predictions = [
-        Y_train_centroid[neighbor_predict]
-        for neighbor_predict in neighbors 
-    ]
-
-    return np.stack(predictions), Y_val
-
-
-# In[18]:
-
-
-knn_model = KNeighborsClassifier(n_neighbors=10)
-
-
-# In[19]:
+# In[16]:
 
 
 _ = cross_validation(
@@ -438,75 +225,40 @@ _ = cross_validation(
 
 # This section I'm using the entire dataset to calculate journal centroids and then evaluate performance on the sub-sampled dataset.
 
+# In[19]:
+
+
+predictions, true_labels = (
+    knn_centroid_evaluate(
+        knn_model,
+        full_training_dataset[300], 
+        subsampled_training_dataset[300]
+    )
+)
+
+
 # In[20]:
 
 
-def knn_centroid_full_evaluate(model, training_data, validation_data, **kwargs):
-    
-    train_centroid_df = (
-        kwargs.get("full_dataset")
-        .groupby("journal")
-        .agg("mean")
-        .reset_index()
-    )
-            
-    X_train_centroid = (
-        train_centroid_df
-        .drop("journal", axis=1)
-        .values
-        .astype('float32')
-    )
-
-    Y_train_centroid = (
-        train_centroid_df
-        .journal
-        .values
-    )
-    
-    
-    X_val = (
-        validation_data
-        .drop("journal", axis=1)
-        .values
-        .astype('float32')
-    )
-    
-    Y_val = (
-        validation_data
-        .journal
-        .values
-    )
-    
-    knn_model.fit(X_train_centroid, Y_train_centroid)
-    distance, neighbors = knn_model.kneighbors(X_val)
-    
-    predictions = [
-        Y_train_centroid[neighbor_predict]
-        for neighbor_predict in neighbors 
-    ]
-
-    return np.stack(predictions), Y_val
+accs = [
+    (
+         1 if true_labels[data_idx] in prediction_row 
+         else 0 
+     )
+     for data_idx, prediction_row in enumerate(predictions)
+]
 
 
 # In[21]:
 
 
-knn_model = KNeighborsClassifier(n_neighbors=10)
-
-
-# In[23]:
-
-
-_ = cross_validation(
-    knn_model, subsampled_training_dataset[300], 
-    knn_centroid_full_evaluate, cv=10, 
-    random_state=100, full_dataset=full_training_dataset[300]
-)
+print(f"{np.sum(accs)} out of {len(accs)}")
+print(f"{np.mean(accs)}% correct")
 
 
 # # Golden Set Analysis
 
-# In[25]:
+# In[22]:
 
 
 biorxiv_embeddings_df = pd.read_csv(
@@ -518,7 +270,7 @@ biorxiv_embeddings_df = pd.read_csv(
 biorxiv_embeddings_df.head()
 
 
-# In[35]:
+# In[23]:
 
 
 golden_dataset = (
@@ -531,60 +283,31 @@ golden_dataset.head()
 
 # ## Centroid Analysis
 
-# In[37]:
+# In[25]:
 
 
-train_centroid_df = (
-    full_training_dataset[300]
-    .groupby("journal")
-    .agg("mean")
-    .reset_index()
-)
-
-X_train_centroid = (
-    train_centroid_df
-    .drop("journal", axis=1)
-    .values
-    .astype('float32')
-)
-
-Y_train_centroid = (
-    train_centroid_df
-    .journal
-    .values
-)
-knn_model = KNeighborsClassifier(n_neighbors=10)
-knn_model.fit(X_train_centroid, Y_train_centroid)
-
-
-# In[38]:
-
-
-distance, neighbors = knn_model.kneighbors(
-    golden_dataset
-    .drop(["document", "pmcid", "journal"], axis=1)
-    .values
+predictions, true_labels = (
+    knn_centroid_evaluate(
+        knn_model,
+        full_training_dataset[300], 
+        golden_dataset.drop(["pmcid", "document"], axis=1)
+    )
 )
 
 
-# In[40]:
+# In[26]:
 
 
 accs = [
     (
-         1 if golden_dataset.journal[data_idx] in prediction_row 
+         1 if true_labels[data_idx] in prediction_row 
          else 0 
      )
-     for data_idx, prediction_row in enumerate(
-         [
-             Y_train_centroid[neighbor_predict]
-             for neighbor_predict in neighbors 
-         ]
-     )
+     for data_idx, prediction_row in enumerate(predictions)
 ]
 
 
-# In[43]:
+# In[27]:
 
 
 print(f"{np.sum(accs)} out of {len(accs)}")
@@ -593,53 +316,31 @@ print(f"{np.mean(accs)}% correct")
 
 # ## Subsampled Paper Analysis
 
-# In[44]:
+# In[28]:
 
 
-X_train = (
-    subsampled_training_dataset[300]
-    .drop("journal", axis=1)
-    .values
-    .astype('float32')
-)
-    
-Y_train = (
-    subsampled_training_dataset[300]
-    .journal
-    .values
-)
-knn_model = KNeighborsClassifier(n_neighbors=10)
-knn_model.fit(X_train, Y_train)
-
-
-# In[45]:
-
-
-distance, neighbors = knn_model.kneighbors(
-    golden_dataset
-    .drop(["document", "pmcid", "journal"], axis=1)
-    .values
+predictions, true_labels = (
+    knn_evaluate(
+        knn_model,
+        subsampled_training_dataset[300],
+        golden_dataset.drop(["pmcid", "document"], axis=1)
+    )
 )
 
 
-# In[46]:
+# In[29]:
 
 
 accs = [
     (
-         1 if golden_dataset.journal[data_idx] in prediction_row 
+         1 if true_labels[data_idx] in prediction_row 
          else 0 
      )
-     for data_idx, prediction_row in enumerate(
-         [
-             Y_train[neighbor_predict]
-             for neighbor_predict in neighbors 
-         ]
-     )
+     for data_idx, prediction_row in enumerate(predictions)
 ]
 
 
-# In[47]:
+# In[30]:
 
 
 print(f"{np.sum(accs)} out of {len(accs)}")
