+++ setup.py
@@ -1,0 +1,9 @@
+from setuptools import setup
+
+setup(
+    name='annorxiver_modules',
+    author='David Nicholson',
+    version='0.0.1',
+    description='Helper functions for the annorxiver project',
+    packages=['annorxiver_modules']
+)
